Return a list instead of set
Library and Relocation don't override equals and hashCode, so using a Set is useless in this case and can be replaced with a List. Also, in the rest of the library a List is commonly used when handling such types.
A List is used also for ExcludedDependency to keep the API more uniform.